v2 phase 0, commit 1: extract the scan core into src/scan/ (pure motion)
The mode-agnostic scan machinery moves out of src/backfill/ into
src/scan/: process-batch, persist, routing, batch-types, output, timing,
util. src/backfill/ shrinks to the day-walk driver (run, args, days,
discovery, entity, structure-mode, index) — "which files + what date";
the core is "what happens to a batch". Forward-sync-v2 plugs in as a
second driver feeding the same core.

Behavior-preserving: file moves + import-path rewrites only. entity.ts
stays driver-side (loadEntity is Args/cursor/cache-flavored); the scan
core temporarily type-imports EntityCtx from it — the type moves into
scan in the next commit.
