+++ src/scan/batch-types.ts
