+++ src/scan/output.ts
