+++ src/scan/routing.ts
