+++ src/scan/timing.ts
