+++ src/scan/util.ts
